% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Cube a vector}
\usage{
cube(x)
}
\arguments{
\item{x}{The vector to be cubed.}
}
\value{
A vector that is the cube of \code{x}.
}
\description{
This function simply cubes a vector!
}
\details{
A number multiplied thrice by itself is called "cubed" or "to the power of 3."

This function utilizes the pow() function in the same package to calculate cubes of numbers.

This function also returns an informative error when a non-numeric input is given.
}
\examples{
cube(1:10)
cube(-3)
}
